+++ src/world/astronomy/orbit.rs
@@ -71,12 +71,12 @@ pub fn orbital_positioning_system(
                 parent_mass,
                 game_time.elapsed_secs(),
             );
-            transform.translation = parent_translation + (Vec3::from(pos) / 2_000_000.0);
+            transform.translation =
+                parent_translation + (Vec3::from(pos) / 10_000_000.0).clamp_length_min(250.0);
             // TODO: Maybe use a different scaling factor depending on the _kind_ of
             //       body being orbited (star, planet, etc.).
-            // NOTE: You can also .clamp_length() of a Vec3 so that it doesn't exceed
-            //       or fall below some value(s). Useful for ensuring an orbiting body
-            //       does not appear inside of another one.
+            // NOTE: .clamp_length_min() ensures that the Moon and/or other close satellites
+            //       do not fall below a certain distance from their parent orbitable.
         }
     }
 }