+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Decrease static orbit scale (planets are now closer)
 - Clamp orbit distance (prevents Moon orbiting inside of Earth sprite)
+- Indicators spawn on update allowing new entities to be pointed to
 
 ## [0.0.25] - 2023-11-17